pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentgovbench"
version = "0.2.0"
description = "An open benchmark for AI agent governance. Mapped to NIST AI RMF. Vendor-neutral."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
  { name = "AgentGovBench contributors" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Security",
]
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31",
  "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[project.scripts]
agentgovbench = "benchmark.cli:main"

[tool.setuptools.packages.find]
include = ["benchmark*", "runners*"]
